Rental flow nearly done - missing RideShare

+++ Car Rental Service/CarApp+Gateway/Shared/Dto/PaymentRequestDto.cs
@@ -2,10 +2,10 @@
 {
     public class PaymentRequestDto
     {
-        public long Customer { get; set; } // Maps to int64
+        public int Customer { get; set; } // Maps to int64
         public int BookingType { get; set; } // Maps to int32
-        public long Booking { get; set; } // Maps to int64
+        public int Booking { get; set; } // Maps to int64
         public string Status { get; set; } // Maps to string
-        public long CreditCardRef { get; set; } // Maps to int64
+        public int CreditCardRef { get; set; } // Maps to int64
     }
 }
